// src/app/shared/Model/Catalogue.ts
export class Catalogue {
  name: string;
  description: string;
  categories: string;
  enseignes: string;
  commercant: string;
  typecompagne: string;
  dateDebut: string;
  heureDebut: string;
  dateFin: string;
  heureFin: string;
  pdf: string;
  cover: string;
  etat: string;
  tags: any;
  constructor() { }
  // public set $commercant(x: Commercant) {
  //     this.commercant = x
  // }
}